refactor: 移動した埋め込みとボタン作成関数のインポートパスを修正し、共通の埋め込みファイルを作成
