+++ src/commands/queue/rank/vote.ts
@@ -3,7 +3,7 @@ import { MessageFlags } from 'discord.js'
 import type { LolRole } from '@/constants'
 import { logger } from '@/lib/logger'
 import { apiClient } from '@/utils/api-client'
-import { createMatchEmbed, createMatchResultEmbed, createVoteButtons } from '../shared/embeds'
+import { createMatchEmbed, createMatchResultEmbed, createVoteButtons } from './embeds'
 import type { TeamAssignments } from '../shared/types'
 
 export const handleVote = async (interaction: ButtonInteraction<CacheType>, matchId: string, vote: 'blue' | 'red') => {
